edit absence

+++ database/migrations/2022_06_03_094218_create_stagiaires_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string("prenom_st");
             $table->foreignId("groupe_id");
             $table->string("status")->default("Active");
-            $table->integer("heure_absence_st")->default(0);
+            $table->float("heure_absence_st")->default(0);
             $table->string("numero_personnelle")->default("-----");
         });
     }
